Starting to declare syntactic sugar factory methods in domain-specific classes.
The naming convention I've adopted is MatchXxxs, inspired by Colin Vipurs' idea on the mailing list of how to help people find matchers with autocomplete.  E.g. MatchStrings for string matchers, MatchIterables for iterable matchers, and so on.

The factory classes are typically (but not always) in the same subpackage of the matchers that they provide.  E.g. MatchStrings is in org.hamcrest.text, MatchBeans in org.hamcrest.beans.  The core matchers fall do not fit this scheme.  I'm not yet sure what to do about allOf, anyOf, not, equalTo, anything(), etc..  For now, equalTo and anything() are on MatchObjects but maybe the core matchers should remain on Matchers, and all the other factory methods be deprecated,

 I haven't deprecated any of the existing factory methods on the Matcher subclasses yet.